added inventory list and purchase list components
