+++ src/components/Dashboard/Dashboard.jsx
@@ -1,25 +1,9 @@
-import { useContext, useEffect, useState } from 'react';
-
+import { useContext} from 'react';
 import { UserContext } from '../../contexts/UserContext';
-
-import * as listService from "../../services/listService";
 
 const Dashboard = (props) => {
 
-  const { user } = useContext(UserContext);
-
-  const [lists, setLists] = useState([]);
-
-  useEffect(() => {
-    const fetchAllLists = async () => {
-      const listsData = await listService.index();
-  
-      setLists(listsData)
-    
-    };
-    if (user) fetchAllLists();
-  }, [user]);
-  
+  const { user } = useContext(UserContext); 
 
   return (
     <main>
@@ -27,7 +11,7 @@ const Dashboard = (props) => {
       <p>
         This is the dashboard page where you can see all lists.
       </p>
-      {lists.map((list) => (
+      {props.lists.map((list) => (
         <p key={list._id}>{list.title}</p>
       ))}
     </main>
