+++ src/components/ListForm/ListForm.jsx
@@ -24,7 +24,7 @@ const ListForm = (props) => {
         <label htmlFor='listType'>List Type</label>
         <select required name='listType' id='listType' value={formData.listType} onChange={handleChange}
 >
-          <option value='Inventory'>Inventory</option>
+          <option value='Inventory'>Inventory list</option>
           <option value='Purchase list'>Purchase List</option>
 
         </select>
