+++ src/components/PurchaseLists/PurchaseLists.jsx
@@ -1,0 +1,8 @@
+const PurchaseLists = (props) => {
+
+    return (
+        <h1>purchase</h1>
+    );
+};
+
+export default PurchaseLists;
